requirements.txt
wheel
flask
twilio
vicses.viper
boto3
zappa
